+++ src/pages/api/users/[userId].ts
@@ -1,41 +1,25 @@
 import {NextApiRequest, NextApiResponse} from "next";
-import {PrismaClient} from "@prisma/client";
+import setProfile from "@/controller/users";
 
-export default async function userHandler(req: NextApiRequest, res: NextApiResponse) {
-    const {query, method, body} = req;
-    // const id = parseInt(query.userId as string, 10)
-    console.log(req.body)
+export default function userHandler(req: NextApiRequest, res: NextApiResponse) {
+    const {query, method} = req;
+    const id = parseInt(query.userId as string, 10)
+    const name = query.name as string
 
     switch (method) {
         case 'GET':
             // Get data from your database
-            // res.status(200).json({ id, name: `User ${id}` })
+            res.status(200).json({ id, name: `User ${id}` })
             break
         case 'POST':
-            try {
-                const prisma = new PrismaClient()
-                const response = await prisma.user_account.create({
-                    data: {
-                        id: req.body.uid,
-                        name: req.body.name,
-                        introduce: req.body.introduce,
-                        score: 0,
-                        status: 'login'
-                    },
-                })
-                console.log(response)
-            } catch (e) {
-                console.log(e)
-            }
-
             // console.log(`${req.body.name}`)
-            // setProfile(body)
+            setProfile(req.body)
             // Get data from your database
             // res.status(200).json({ id, name: `User ${id}` })
             break
         case 'PUT':
             // Update or create data in your database
-            // res.status(200).json({ id, name: name || `User ${id}` })
+            res.status(200).json({ id, name: name || `User ${id}` })
             break
         default:
             res.setHeader('Allow', ['GET', 'PUT'])
